Fix errors when using upcoming scala 3.5.0
Previously, a bug in the compiler could remove some parts of the code
before an inline match. In this project, this prevented the exception
`inline val with null is not supported` from appearing, since this check
is done later in the compilation pipeline. Since that is fixed in 3.5.0,
the exception can appear now.

Unfortunately, currently I do not have a better solution for this.
`case res => res` (for now) only works in scala 3.4.0 and later.